pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PI-R"
version = "1.0.0"
description = "PI-R"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
   "packaging==24.2",
   "wandb==0.19.8",
   "deepspeed==0.16.3",
   "accelerate==0.30.1",
   "diffusers==0.32.2",
   "timm==1.0.3",
   "transformers==4.48.1",
   "sentencepiece==0.2.0",
   "h5py==3.11.0",
   "opencv-python==4.9.0.80",
   "imgaug==0.4.0",
   "huggingface_hub==0.29.3",
   "numpy==1.26.4",
   "tokenizers==0.21.1",
   "safetensors==0.4.3",
   "termcolor==2.5.0",
   "draccus==0.10.0",
   "bitsandbytes==0.42.0",
   "datasets==3.3.2",
   "jsonlines==4.0.0",
   "pytest==8.3.5",
   "tyro==0.9.5",
   "websockets==14.1",
   "hydra-core==1.3.2",
   "omegaconf==2.3.0",
   "colorama==0.4.6"
]


[tool.setuptools.packages.find]
exclude = []

[tool.wheel]
exclude = []
